created a toggle button for asc and desc sorting, added states with toggleSortDirection and handleSortFieldChange functions, separated components in their folders
