+++ src/assets/components/AddTodoForm/AddTodoForm.module.css
@@ -1,0 +1,6 @@
+.inputForm{
+    display: flex;
+    align-items: center;
+    font-weight: bold;
+    gap: 10px; 
+}
