+++ src/assets/pages/Home/Home.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import styles from '../pages/Home.module.css'
+import styles from '../Home/Home.module.css'
 
 const Home = () => {
     const navigate = useNavigate();
